Updating paged fronthistory results to truncate long member lists
Incorporates xSke's fix for fronthistory embed descriptions exceeding the limit by setting a per-line length restriction based on the page size. Making the page size dynamic from input was straightforward, so I've done that as well - that way if your member list is getting truncated you can at least try a smaller page size. Page size defaults to 10 (current behavior) if it's not specified, and the allowable range is from 1 to 25.