fix(bot): fix import/export with incorrect webhook_avatar_url
Incorrect use of `member.WebhookAvatarFor` in member JSON creation
meant that exports of members without a webhook avatar set had the
normal avatar URL present in both fields.

This also adds a check to the importer to ignore the value of the
webhook_avatar_url field if it is the same as the avatar_url field.

Fixes: ccb89f50e96861d5c49e7a31277d3f0bf1b3afba (feat(bot): allow separate member avatars for proxied messages)
Co-authored-by: spiral <spiral@spiral.sh>